Code Quality Commit
