+++ readme.md
@@ -1,4 +1,6 @@
-# Parent library for all Codeception framework modules and PhpBrowser
+# Codeception Lib InnerBrowser
+
+Parent library for all Codeception framework modules and PhpBrowser.
 
 [![Build Status](https://travis-ci.org/Codeception/lib-innerbrowser.svg?branch=master)](https://travis-ci.org/Codeception/lib-innerbrowser)
